Support single/double precision.
In order to support single/double precision, I changed the real types to void*
so that they can be type cast to the proper type. This is safe as long as the
code only uses the bml_matrix_t type since the precision is stored inside it.
The application has to ensure though that the conversion functions are used
properly, i.e. that the right real type is used for the arguments or the
return type.

On the other hand, this approach does not require a preprocessor based
double compilation of the code for the two real types, and a duplication of
the APIs and function names.