Fix bandwidth test
Since Fortran and C agree to disagree on dense matrix element ordering, we
really needed a way to tell the library which order a dense matrix was in
(like the string argument "N" or "T" in BLAS type APIs). I have added those
and now the bandwidth works properly across the Fortran interface.