Fixes Issue #2
Added valgrind tests for Fortran APIs. Also fixed a memory leak in the Fortran
copy operation, in which an array assignment causes an implicit allocation on
the lhs. The added deallocation of the rhs will likely break if the compiler
is told to not implicitly allocate the lhs on assignment. This needs to get
addressed then.