Also test for docs and formatting
Added a test that builds the doxygen documentation with the ultimate goal of
flagging _missing_ documentation as a build error. At this point, this test
only emits warnings.

Added a test for code formatting. The test runs the indent.sh script which
applies the project's coding style to all sources and then diffs the new
sources against the old ones. If any differences are found, this test fails.