Add static library tests to Travis-CI
During the testing of this PR I built the code with static libraries and
found that everything is passing. Travis-CI however is building only
shared libraries and found that the PR is _not_ building and the library
is missing some symbols. We should test both static and shared libraries
for this reason. It should be sufficient to test one compiler version
with static libraries since libraries are managed by ld and not the
compiler.