Relax tolerance for inverse test
The inverse-dense-single_real test is failing for GCC 6 and OpenMP:

Error in matrix inverse; ssum(A*A_inverse) = -4.949570e-04

The test calculates the inverse, takes the product of the original
matrix and this inverse, and then computes the difference between N and
the sum of the squares of all matrix elements. In addition, with OpenMP
the implementation of the reduction affects the final result.

The final result, `ssum`, is then the accuracy of the inverse, the
matrix product `A*A^{-1}`, and the summed **squares** of the differences
of all elements. It's unclear to me exactly what accuracy to expect, but
the one we are using is not achieved for GCC 6 and OpenMP.