Added example to replace min/max reductions
The min/max reductions in C/C++ were added to the OpenMP standard in
OpenMP 3.1 which is available in >=gcc-4.7. In order to support earlier
compilers we have to replace all occurrences of min/max reduction
clauses with an appropriate alternative. This change adds an example of
such an alternative. The example implements the reduction using the max
reduction clause and the workaround.

    $ time ./min_max_reduction -N 1000000000
    using < OpenMP-3.1
    running on 8 threads
    emax = 999999999

    real        0m0.006s
    user        0m0.002s
    sys         0m0.007s

    $ time ./min_max_reduction -N 1000000000 -n
    using >= OpenMP-3.1
    running on 8 threads
    emax = 999999999

    real        0m0.010s
    user        0m0.031s
    sys         0m0.006s

Within the granularity of the 'time' command, there is no performance
difference between the two implementations.