Reduce the number of compilers tested
We have not seen any issues with the intermediate compiler versions in
the Travis-CI tests. This change reduces the tests to `gcc-4.7`, the
first `gcc` compiler to support the `OpenMP-3.1` standard and the most
recent compiler in the `ppa:ubuntu-toolchain-r/test` repository,
`gcc-6` in order to speed up the test runs.