Print out coverty log files
For debugging coverty issues it could be useful to know the contents of
the coverty log files. Since storing of build artifacts requires a more
complicated Travis-CI setup with S3 backed storage, a simple work around
is to print the contents of the log files so that their content is
visible in the build logs.