Restrict coverty scan with Travis-CI build matrix
When running coverty scan the Travis-CI build matrix causes the scan to
run for every element of the matrix. This is unnecessary. This change
leverages the environment variables `COVERITY_SCAN_BRANCH` and
`TRAVIS_JOB_NUMBER` to determine whether a coverty scan is run and to
restrict a coverty scan to the first element of the build matrix.