Add new format 'ellblock'
This is based on the ellpack format, but each "element" is now a block.
Block sizes are defined either by the first call to build a matrix with
'bml_block_matrix' or by their default values. They have to be the same
for all matrices. This code seems to work properly and pass all the tests.
It has however not been optimized and does not include OpenMP directives.