Make build of complex types optional
The Nvidia compilers don't support complex types for device. This
change adds an option to disable building of complex types. In
addition CMake will now verify that the compile can deal with complex
types.

This implementation is pretty crude. If complex types are not built,
then the corresponding types will be mapped to their real valued
counter parts in `typed.h`. The library will still contain functions
that have a `complex` in their name.

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>