Solved a bug in bml_export_to_dense with magma
When bml_expoert_to_dense was invoked and bml was compiled with magma,
the original bml matrix was modified (transposed). We are adding
another "back" transpose to go back to the original matrix. Another
solution would be to allocate an auxiliary matrix and let the original
unchanged, but this implies the use of extra memory.