Re-run skipped tests
Skipped tests do not produce a failure and no output with the
`--output-on-failure` command line options. This makes debugging
inconvenient. We can not simple run the whole test suite with
`--verbose` because this would produce too much output and Travis-CI
will terminate the test. This change re-runs the skipped tests with
the `--verbose` argument.

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>