Do not use pre-built images for CI
This will allow us to modify the build environment from git since we
don't depend on our own images anymore.

Note, the only reason we might want to use pre-built images is
turnaround time for CI. But caching on the CI infrastructure should
make the building of the container pretty fast.

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>