Add testing of remaining float types to MPI tests
So far we are testing only the `single_real` float type for the MPI
tests. This change adds the other float types.

This change disables `valgrind` testing for the MPI jobs. Valgrind
leads to many false positives under MPI because of memory issue within
MPI itself. In order to enable valgrind for MPI again we need to
exclude the MPI related libraries.

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>