Build CI environment in workflow
This change removes the custom built container from the GitHub
workflow and instead configures the build environment during the CI
run itself.

This has the advantage that anyone (including forks) can modify the
build environment in a pull request and use it for that same pull
request in CI. In other words a change in the CI environment itself,
such as a compiler upgrade, can be tested directly in the pull
request.

In addition, environment secrets [1] for pushing the custom built
Docker container to Docker Hub are now not necessary anymore, reducing
the attack surface of the project.

[1] https://docs.github.com/en/actions/reference/encrypted-secrets

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>