Build image on `push` event
After a successful code review via a pull request we need to rebuild
the container image in `master` so that the image is updated on Docker
Hub.

For security, the current conditional in the workflow definition
allows such image builds only on branches in the `lanl/bml` repository
for pull requests. Once such a pull request is merged we need to
rebuild the image on the `master` branch so that it is tagged
correctly on Docker Hub.

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>