Add pre-commit hook script
The pre-commit hook script needs to be copied locally to:

  $ cp scripts/pre-commit .git/hooks/pre-commit

It will run the linter before committing.

Since the linter takes a while to complete, it can be helpful to disable
this check if desired. Adding the `--no-verify` command line option to
`git commit` will do that:

  $ git commit --no-verify [options]

Note that for reasons of reproducibility the linter runs inside a docker
container. Therefore, the only pre-requisite for the hook is docker.

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>